+++ tests/functional/java/tests/unit/WaveFileAudioInputStream.java
@@ -38,7 +38,7 @@ public class WaveFileAudioInputStream extends AudioInputStream {
      * @return the number of bytes have been read.
      */
     @Override
-    public int read(byte[] dataBuffer) {
+    public long read(byte[] dataBuffer) {
         if(dataBuffer == null) throw new NullPointerException("dataBuffer");
         
         try {
